# contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Egypt Industrial Initiatives Platform API",
  "version": "1.1.0"
 },
 "servers": [
  {
   "url": "http://localhost:4000"
  },
  {
   "url": "https://api.industry.gov.eg"
  }
 ],
 "paths": {
...
  "/api/v1/applications/{id}/decisions": {
   "post": {
    "summary": "Decide the current stage (only the organization that owns it; officials decide the ministry intake stage only)",
    "responses": {
     "200": {
      "description": "decision recorded"
     },
     "400": {
      "description": "VALIDATION_ERROR | COMMENTS_REQUIRED | ACTION_NOT_ALLOWED"
     },
     "403": {
      "description": "NOT_ASSIGNED (other organization, officials monitoring, auditor)"
     },
     "404": {
      "description": "not found"
     },
     "409": {
      "description": "APPLICATION_CLOSED"
     },
     "429": {
      "description": "rate limited"
     }
    }
   }
  },
...
 }
}